+++ gui/settingsChanged.h
