+++ ocr/easyocr.cpp
