+++ ocr/ocrfunctions.cpp
