+++ ocr/readingItems.cpp
