+++ ocr/tesseract.cpp
