+++ relics/relicDatabase.cpp
